requirements/development.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file=requirements/development.txt requirements/development.in
#
amqp==5.0.6
    # via kombu
asgiref==3.4.1
    # via
    #   django
    #   djoser
attrs==21.2.0
    # via pytest
billiard==3.6.4.0
    # via celery
celery==5.1.2
    # via -r requirements/base.in
certifi==2021.5.30
    # via requests
cffi==1.14.6
    # via cryptography
charset-normalizer==2.0.3
    # via requests
click-didyoumean==0.0.3
    # via celery
click-plugins==1.1.1
    # via celery
click-repl==0.2.0
    # via celery
click==7.1.2
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
coreapi==2.3.3
    # via djoser
coreschema==0.0.4
    # via coreapi
cryptography==3.4.7
    # via social-auth-core
defusedxml==0.7.1
    # via
    #   python3-openid
    #   social-auth-core
django-cors-headers==3.7.0
    # via -r requirements/base.in
django-templated-mail==1.1.1
    # via djoser
django==3.2.5
    # via
    #   django-cors-headers
    #   djangorestframework
    #   djangorestframework-simplejwt
    #   drf-nested-routers
djangorestframework-simplejwt==4.7.2
    # via
    #   -r requirements/base.in
    #   djoser
djangorestframework==3.12.4
    # via
    #   -r requirements/base.in
    #   djangorestframework-simplejwt
    #   drf-nested-routers
djoser==2.1.0
    # via -r requirements/base.in
drf-nested-routers==0.93.3
    # via -r requirements/base.in
flake8==3.9.2
    # via -r requirements/development.in
idna==3.2
    # via requests
iniconfig==1.1.1
    # via pytest
itypes==1.2.0
    # via coreapi
jinja2==3.0.1
    # via coreschema
kombu==5.1.0
    # via celery
lxml==4.6.3
    # via -r requirements/base.in
markupsafe==2.0.1
    # via jinja2
mccabe==0.6.1
    # via flake8
oauthlib==3.1.1
    # via
    #   requests-oauthlib
    #   social-auth-core
packaging==21.0
    # via pytest
pluggy==0.13.1
    # via pytest
prompt-toolkit==3.0.19
    # via click-repl
py==1.10.0
    # via pytest
pycodestyle==2.7.0
    # via flake8
pycparser==2.20
    # via cffi
pyflakes==2.3.1
    # via flake8
pyjwt==2.1.0
    # via
    #   djangorestframework-simplejwt
    #   social-auth-core
pyparsing==2.4.7
    # via packaging
pytest==6.2.4
    # via -r requirements/development.in
python-dotenv==0.18.0
    # via -r requirements/base.in
python3-openid==3.2.0
    # via social-auth-core
pytz==2021.1
    # via
    #   celery
    #   django
redis==3.5.3
    # via -r requirements/base.in
requests-oauthlib==1.3.0
    # via social-auth-core
requests==2.26.0
    # via
    #   coreapi
    #   requests-oauthlib
    #   social-auth-core
six==1.16.0
    # via
    #   click-repl
    #   social-auth-app-django
social-auth-app-django==4.0.0
    # via djoser
social-auth-core==4.1.0
    # via social-auth-app-django
sqlparse==0.4.1
    # via django
toml==0.10.2
    # via pytest
uritemplate==3.0.1
    # via coreapi
urllib3==1.26.6
    # via requests
vine==5.0.0
    # via
    #   amqp
    #   celery
    #   kombu
wcwidth==0.2.5
    # via prompt-toolkit

# The following packages are considered to be unsafe in a requirements file:
# setuptools
